# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ship RESTful API",
        "description": "Ship RESTful API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get paginated list of documents",
                "description": "Get paginated list of documents",
                "operationId": "cf58d1781992e8292fd9914b24487fdc",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "description": "Set to 'none' to disable pagination. Defaults to 'true'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "none"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items to return per page. Defaults to 10.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Create a new document",
                "operationId": "1c3ad2906b68a58de1f603260ab271c4",
                "requestBody": {
                    "description": "Request body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Document"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "ppkb": {
                                            "type": "integer",
                                            "example": "123456"
                                        },
                                        "service_code": {
                                            "type": "string",
                                            "example": "siklus_pelayanan_air"
                                        },
                                        "status_doc": {
                                            "type": "string",
                                            "example": "nota"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-01T00:00:00+00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-01T00:00:00+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{'ppkb': ['The ppkb field is required.']}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a single document by ID",
                "operationId": "9cf5289c9ad15ddb6da247477998f54b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Data Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Document": {
                "properties": {
                    "ppkb": {
                        "type": "integer",
                        "example": "2020057650"
                    },
                    "service_code": {
                        "type": "string",
                        "enum": [
                            "siklus_pelayanan_air",
                            "siklus_pelayanan_keberangkatan",
                            "siklus_pelayanan_labuh",
                            "siklus_pelayanan_perpanjangan",
                            "siklus_pelayanan_perubahan_kapal",
                            "siklus_pelayanan_pindah",
                            "siklus_pelayanan_tambat"
                        ]
                    },
                    "status_doc": {
                        "type": "string",
                        "enum": [
                            "nota",
                            "cancel_pkk",
                            "dtjk",
                            "pranota"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}